+++ src/Entities/Enums/SuspiciousAccountActivity.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace GlobalPayments\Api\Entities\Enums;
+
+use GlobalPayments\Api\Entities\Enum;
+
+class SuspiciousAccountActivity extends Enum
+{
+    const SUSPICIOUS_ACTIVITY = 'SUSPICIOUS_ACTIVITY';
+    const NO_SUSPICIOUS_ACTIVITY = 'NO_SUSPICIOUS_ACTIVITY';
+}